requirements.txt
streamlit
langchain==0.3.0
langchain-community==0.3.0
langchain-core==0.3.0
langchain-groq==0.2.0
langchain-openai==0.2.0
langchain-text-splitters==0.3.0
faiss-cpu==1.13.2
sentence-transformers
pypdf==5.1.0
python-dotenv==1.0.1
openai==1.50.0
tiktoken==0.12.0
fastapi==0.124.4
pydantic
uvicorn==0.38.0
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.2.1
python-multipart